tools/sdk/vector-sdk/requirements.txt
aiogrpc>=1.4
googleapis-common-protos
opencv-python
numpy
Pillow
requests